pyproject.toml
[build-system]
requires = [
    "setuptools"
]
build-backend = "setuptools.build_meta"

[project]
name = "trading-system-autotest"
version = "0.0.1"
authors = [
    { name = "syl" }
]
description = "Autotest project"

dependencies = [
    "selenium >= 4.11.2",
    "seleniumbase >= 4.18.7",
    "pytest >= 7.4.2",
    "pytest-rerunfailures >= 12.0",
    "pytest-assume >= 2.4.3",
    "webdriver-manager >= 4.0.0",
    "PyYAML >= 6.0.1",
    # "pytest-testmon >=  2.0.12",
    "aircv >= 1.4.6",
    "opencv-python >= 4.8.0",
    "allure-pytest >= 2.13.2",
    "ddddocr >= 1.4.8"
]

[tool.setuptools.packages.find]
exclude = [
    "img",
    "page",
    "logs",
    "base",
    "config",
    "common",
    "testcases",
]
